wss-server/requirements.txt
requests~=2.28.2
websocket-client~=1.5.1
psutil~=5.9.4
Django~=4.1.6
djangorestframework~=3.14.0
asgiref~=3.6.0
channels~=4.0.0